+++ packages/amount-input/src/Component.tsx
@@ -1,13 +1,14 @@
 import React, {
-    type FocusEvent,
     forwardRef,
     Fragment,
     useCallback,
-    useEffect,
     useLayoutEffect,
     useMemo,
+    useRef,
     useState,
 } from 'react';
+import mergeRefs from 'react-merge-refs';
+import { useMaskito } from '@maskito/react';
 import cn from 'classnames';
 
 import { Input, type InputProps } from '@alfalab/core-components-input';
@@ -16,15 +17,15 @@ import { getMinMaxOrDefault } from '@alfalab/core-components-number-input/utils'
 import { fnUtils, isNonNullable } from '@alfalab/core-components-shared';
 import { withSuffix, type withSuffixProps } from '@alfalab/core-components-with-suffix';
 import { type CurrencyCodes } from '@alfalab/data';
-import { formatAmount, MMSP, THINSP } from '@alfalab/utils';
+import { MMSP, THINSP } from '@alfalab/utils';
 
 import {
-    getAmountValueFromStr,
-    getCurrencyCodeWithFormat,
-    getFormattedValue,
-    getVisiblePlaceholder,
-    SEP,
-} from './utils';
+    maskitoOptionsGenerator,
+    type NumberParams,
+    parseNumber,
+    stringifyNumber,
+} from './maskito-utils';
+import { getCurrencyCodeWithFormat, getVisiblePlaceholder, SEP } from './utils';
 
 import defaultColors from './default.module.css';
 import styles from './index.module.css';
@@ -171,8 +172,8 @@ export const AmountInput = forwardRef<HTMLInputElement, AmountInputProps>(
             clear = false,
             onChange,
             onClear,
-            onBlur,
             onFocus,
+            onBlur,
             onKeyDown,
             breakpoint,
             client,
@@ -187,46 +188,45 @@ export const AmountInput = forwardRef<HTMLInputElement, AmountInputProps>(
         ref,
     ) => {
         const controlled = isNonNullable(valueFromProps);
+        const numberParams = useMemo<NumberParams>(() => {
+            const minorityLength = Math.trunc(Math.log10(minority));
+            const integerLength = Math.min(integerLengthFromProps, 15);
+            const maximumFractionDigits = integersOnly ? 0 : minorityLength;
+            const max = Number(
+                `${'9'.repeat(integerLength)}${integersOnly ? '' : '.'}${'9'.repeat(maximumFractionDigits)}`,
+            );
 
+            return {
+                decimalSeparator: SEP,
+                thousandSeparator: MMSP,
+                maximumFractionDigits,
+                minusSign: '-',
+                min: positiveOnly ? 0 : max * -1,
+                max,
+            };
+        }, [integerLengthFromProps, integersOnly, minority, positiveOnly]);
+        const inputRef = useMaskito({ options: maskitoOptionsGenerator(numberParams, view) });
         const getFormattedAmount = useCallback(
-            (val: string | number | null) => {
-                if (val === '' || val === null || val === '-') return '';
-
-                return formatAmount({
-                    value: parseInt(`${val}`, 10),
-                    currency,
-                    minority,
-                    view,
-                    negativeSymbol: 'hyphen-minus',
-                }).formatted;
-            },
-            [currency, minority, view],
+            (val: number | string | null) => stringifyNumber(val, numberParams, view),
+            [numberParams, view],
         );
+        const getFormattedAmountRef = useRef(getFormattedAmount);
         const [numberValue, setNumberValue] = useState(valueFromProps ?? defaultValue);
         const [inputValue, setInputValue] = useState(() => getFormattedAmount(numberValue));
-        const [shouldSyncInputValue, setShouldSyncInputValue] = useState(false);
 
         if (controlled && !(numberValue === valueFromProps)) {
             setNumberValue(valueFromProps);
+            setInputValue(getFormattedAmount(valueFromProps));
         }
 
         useLayoutEffect(() => {
-            setInputValue(getFormattedAmount(numberValue));
-        }, [getFormattedAmount, numberValue]);
-
-        useEffect(() => {
-            if (shouldSyncInputValue) {
+            if (!(getFormattedAmountRef.current === getFormattedAmount)) {
                 setInputValue(getFormattedAmount(numberValue));
-                setShouldSyncInputValue(false);
+                getFormattedAmountRef.current = getFormattedAmount;
             }
-        }, [getFormattedAmount, shouldSyncInputValue, numberValue]);
+        }, [getFormattedAmount, numberValue, valueFromProps]);
 
-        const numberValueOrZero = useMemo(
-            () => parseInt(`${numberValue || '0'}`, 10),
-            [numberValue],
-        );
-
-        const integerLength = Math.min(integerLengthFromProps, 15);
+        const numberValueOrZero = useMemo(() => Number(`${numberValue || '0'}`), [numberValue]);
 
         const { min: minStepperValue, max: maxStepperValue } = getMinMaxOrDefault({
             minProp: stepper?.min,
@@ -234,118 +234,47 @@ export const AmountInput = forwardRef<HTMLInputElement, AmountInputProps>(
         });
         const withStepper = !fnUtils.isNil(stepper?.step);
 
-        const minorityLength = Math.trunc(Math.log10(minority));
-
         const [isFocused, setIsFocused] = useState(false);
-        const [majorPart, minorPart] = inputValue.split(SEP);
+        const [majorPart, minorPart] = inputValue.split(numberParams.decimalSeparator);
         const currencyCode = getCurrencyCodeWithFormat(currency, codeFormat);
 
         const handleChange = (event: React.ChangeEvent<HTMLInputElement>) => {
-            const input = event.target;
-            let caret = input.selectionStart ?? 0;
-            let enteredValue = input.value.replace('.', SEP).replace(/[^0-9,-]/g, '');
+            const { value } = event.target;
 
-            if (integersOnly) {
-                [enteredValue] = enteredValue.split(SEP);
+            setInputValue(value);
+
+            const [, minor = ''] = value.split(numberParams.decimalSeparator);
+
+            if (
+                value === numberParams.minusSign ||
+                value.endsWith(numberParams.decimalSeparator) ||
+                (minor.length > 0 && minor.length < numberParams.maximumFractionDigits)
+            ) {
+                return;
             }
 
-            // Эта проверка нужна для того, чтобы обрабатывать значение, переданное в input, длина которого превышает integerLength
-            const integer = integersOnly ? enteredValue : enteredValue.split(SEP)[0];
+            const numOrNull = parseNumber(value, numberParams);
 
-            if (integer.length > integerLength) {
-                enteredValue = enteredValue.slice(0, integerLength);
+            onChange?.(event, {
+                value: numOrNull,
+                valueString: value,
+            });
+
+            if (!controlled) {
+                setNumberValue(numOrNull);
             }
-
-            // Сокращение минимальной длины мажорной части числа до 0 позволяет ввести "," => "0,"
-            const isCorrectEnteredValue = RegExp(
-                `(^${positiveOnly ? '' : '-?'}[0-9]{0,${integerLength}}(,([0-9]+)?)?$|^\\s*$)`,
-            ).test(enteredValue);
-
-            if (isCorrectEnteredValue) {
-                if (inputValue[caret] === MMSP) {
-                    enteredValue = enteredValue.slice(0, caret - 1) + enteredValue.slice(caret);
-                    caret -= 1;
-                }
-
-                const newFormattedValue = getFormattedValue(enteredValue, currency, minority);
-
-                if (newFormattedValue === inputValue) {
-                    window.requestAnimationFrame(() => {
-                        input.selectionStart = caret;
-                        input.selectionEnd = caret;
-                    });
-                } else {
-                    /**
-                     * Поддержка положения каретки
-                     * Поскольку при форматировании введенного значения могут появляться символы типа пробела
-                     * или запятая - каретка прыгает в конец и ее необходимо ставить в правильное место
-                     */
-
-                    // Узнаем длину оригинального инпута с условием обрезания лишних символов
-
-                    const [head, tail] = input.value.split(SEP);
-                    let notFormattedEnteredValueLength = head.length;
-
-                    if (tail) {
-                        notFormattedEnteredValueLength += 1; // запятая или точка
-                        notFormattedEnteredValueLength += tail.slice(0, minorityLength).length; // символы в минорной части
-                    }
-
-                    const diff = newFormattedValue.length - notFormattedEnteredValueLength;
-
-                    caret += diff;
-
-                    window.requestAnimationFrame(() => {
-                        input.selectionStart = caret;
-                        input.selectionEnd = caret;
-                    });
-                }
-
-                setInputValue(newFormattedValue);
-
-                const [, minor = ''] = newFormattedValue.split(SEP);
-
-                if (!integersOnly && minor.replace(/[^0-9]/g, '').length < minorityLength) {
-                    return;
-                }
-
-                const nextValue = getAmountValueFromStr(newFormattedValue, minority);
-
-                onChange?.(event, {
-                    value: nextValue,
-                    valueString: newFormattedValue,
-                });
-
-                setShouldSyncInputValue(true);
-
-                if (!controlled) {
-                    setNumberValue(nextValue);
-                }
-            } else {
-                caret = -1;
-
-                window.requestAnimationFrame(() => {
-                    input.selectionStart = caret;
-                    input.selectionEnd = caret;
-                });
-            }
-        };
-
-        const handleKeyDown = (event: React.KeyboardEvent<HTMLInputElement>) => {
-            const isModifierKeys = event.ctrlKey || event.altKey || event.metaKey || event.shiftKey;
-
-            // Не двигаем каретку когда вводится невалидный символ
-            if (!isModifierKeys && event.key.length === 1 && /[^0-9,.-]/.test(event.key)) {
-                event.preventDefault();
-            }
-
-            onKeyDown?.(event);
         };
 
         const handleFocus: withSuffixProps['onFocus'] = (e) => {
             setIsFocused(true);
 
             onFocus?.(e);
+        };
+
+        const handleBlur: withSuffixProps['onBlur'] = (e) => {
+            setIsFocused(false);
+
+            onBlur?.(e);
         };
 
         const handleClear = useCallback(
@@ -358,62 +287,6 @@ export const AmountInput = forwardRef<HTMLInputElement, AmountInputProps>(
             },
             [controlled, onClear],
         );
-
-        /**
-         * Отбросить десятичный разделитель если находится в конце числа
-         * 123, => 123
-         */
-        const dropDecimalSeparator = (event: FocusEvent<HTMLInputElement>) => {
-            const { value } = event.target;
-
-            const newValue = getAmountValueFromStr(
-                value.endsWith(SEP) ? value.slice(0, -1) : value,
-                minority,
-            );
-
-            const formatted = getFormattedAmount(newValue);
-
-            onChange?.(null, {
-                value: newValue,
-                valueString: formatted,
-            });
-
-            setShouldSyncInputValue(true);
-
-            if (!controlled) {
-                setNumberValue(newValue);
-            }
-        };
-
-        const handleBlur = (event: FocusEvent<HTMLInputElement>) => {
-            const { value } = event.target;
-
-            if (view === 'withZeroMinorPart') {
-                const newValue = getAmountValueFromStr(value, minority);
-                const formatted = getFormattedAmount(newValue);
-
-                if (numberValue !== newValue) {
-                    onChange?.(event, {
-                        value: newValue,
-                        valueString: formatted,
-                    });
-
-                    setShouldSyncInputValue(true);
-
-                    if (!controlled) {
-                        setNumberValue(newValue);
-                    }
-                }
-            }
-
-            if (view === 'default') {
-                dropDecimalSeparator(event);
-            }
-
-            setIsFocused(false);
-
-            onBlur?.(event);
-        };
 
         const handleDecrement = () => {
             if (stepper?.step) {
@@ -476,6 +349,7 @@ export const AmountInput = forwardRef<HTMLInputElement, AmountInputProps>(
                 className={cn(styles.container, {
                     [styles.bold]: bold,
                     [styles.filled]: Boolean(inputValue),
+                    [styles.focused]: isFocused,
                 })}
             >
                 <SuffixInput
@@ -483,7 +357,7 @@ export const AmountInput = forwardRef<HTMLInputElement, AmountInputProps>(
                     rightAddons={renderRightAddons()}
                     suffix={
                         <Fragment>
-                            {majorPart}
+                            <span className={styles.suffixMajor}>{majorPart}</span>
 
                             <span
                                 className={cn({
@@ -492,9 +366,15 @@ export const AmountInput = forwardRef<HTMLInputElement, AmountInputProps>(
                                     [colorStyles[colors].readOnly]: restProps.readOnly,
                                 })}
                             >
-                                {minorPart !== undefined && `${SEP}${minorPart}`}
+                                {minorPart !== undefined && (
+                                    <span
+                                        className={styles.suffixMinor}
+                                    >{`${numberParams.decimalSeparator}${minorPart}`}</span>
+                                )}
                                 {THINSP}
-                                {suffix === currency ? currencyCode : suffix}
+                                <span className={styles.suffixCurrency}>
+                                    {suffix === currency ? currencyCode : suffix}
+                                </span>
                             </span>
                         </Fragment>
                     }
@@ -510,13 +390,11 @@ export const AmountInput = forwardRef<HTMLInputElement, AmountInputProps>(
                     inputClassName={cn(styles.input, inputClassName)}
                     onChange={handleChange}
                     onClear={handleClear}
-                    onBlur={handleBlur}
-                    onKeyDown={handleKeyDown}
                     onFocus={handleFocus}
-                    inputMode='decimal'
-                    pattern={`[${positiveOnly ? '' : '\\-'}0-9\\s\\.,]*`}
+                    onBlur={handleBlur}
+                    inputMode={numberParams.maximumFractionDigits === 0 ? 'numeric' : 'decimal'}
                     dataTestId={dataTestId}
-                    ref={ref}
+                    ref={mergeRefs([ref, inputRef])}
                     breakpoint={breakpoint}
                     client={client}
                 />